+++ EmpiriaBMS.Front/ViewModel/Components/OtherVM.cs
@@ -18,8 +18,8 @@ public class OtherVM : BaseVM
         }
     }
 
-    private double _menHours;
-    public double MenHours
+    private long _menHours;
+    public long MenHours
     {
         get => _menHours;
         set
