+++ EmpiriaBMS.Models/Models/Draw.cs
@@ -15,7 +15,7 @@ public class Draw : Entity
     [Required]
     public string Name { get; set; }
 
-    public double MenHours { get; set; }
+    public long MenHours { get; set; }
 
     public float CompletionEstimation { get; set; }
